New transaction - PayingForTx
Allows an account to pay for a transaction on behalf of someone else.
Paying for fees and gas cost, will not cover the amount spent by the
transaction, just "the cost of the transaction" (and the extra size
added by wrapping the original transaction).

Note: this changes how "gas" for transaction size is calculated for GAs
from IRIS onwards.